design: border-radius 수정 및 색 첨가
