+++ src/components/home-address/Header.tsx
@@ -1,11 +1,13 @@
 import styled from "@emotion/styled";
 import Image from "next/image";
+import { colors } from "@/src/constants/css";
+import { ColorProps } from "@/src/types/css-props";
 
 const HEADER_TEXT = "어디서 만날까?";
 
 const Header = () => {
   return (
-    <HeaderContainer>
+    <HeaderContainer colors={colors}>
       <TextP>{HEADER_TEXT}</TextP>
       <Image
         src='/Logo.png'
@@ -20,13 +22,14 @@ const Header = () => {
 
 export default Header;
 
-const HeaderContainer = styled.header`
+const HeaderContainer = styled.header<{ colors: ColorProps }>`
   min-height: 174px;
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
   color: white;
+  background-color: ${({ colors }) => colors.greenPrimary};
 `;
 
 const TextP = styled.p`
